big file structure refactoring
