+++ gui/customertab.py
@@ -1,0 +1,6 @@
+import wx
+
+
+class CustomerTab(wx.Panel):
+    def __init__(self, parent):
+        super(CustomerTab, self).__init__(parent)
