+++ gui/locale/rus.py
@@ -1,0 +1,25 @@
+APP_NAME = 'Shop UI'
+
+# menubar labels
+HELP = 'Помощь'
+
+# menubar item labels
+ABOUT_ITEM = 'О {}'.format(APP_NAME)
+
+# toolbar item labels
+SET = 'Настройки'
+
+# dialog messages
+ABOUT_DIAL = 'Программа для учёта товаров сферического магазина в вакууме'
+
+# tab labels
+SHOP_TAB = 'Магазин'
+CUSTOMER_TAB = 'Покупатели'
+
+# list column labels
+COL = ['ID', 'Название товара', 'Количество', 'Цена', 'Зарезервировано']
+
+# button labels
+ADD_BUTTON = 'Добавить наименование'
+DELETE_BUTTON = 'Удалить'
+TO_CART_BUTTON = 'Добавить в корзину'
