+++ main.py
@@ -1,0 +1,14 @@
+import wx
+
+import dbmanager
+import gui.mainframe
+
+
+def main():
+    app = wx.App(False)
+    gui.mainframe.MainFrame(None, 'gui/icon/')
+    app.MainLoop()
+
+
+if __name__ == '__main__':
+    main()
