+++ test.py
@@ -1,15 +1,16 @@
 from dbmanager import dbManager
 from dbmanager import sqlite3Adapter
 
-sql3Adpt = sqlite3Adapter.sqlite3Adapter("db/test.db", "id") 
+sql3Adpt = sqlite3Adapter.Sqlite3Adapter("db/test.db", "id")
 db = dbManager.dbManager(sql3Adpt)
 	
 
 # Create table
-sql3Adpt.sqliteCursor.execute('''CREATE TABLE stocks
-             (ID INTEGER PRIMARY KEY AUTOINCREMENT,name varchar(100), count INT, price INT, reserved INT DEFAULT 0)''')
+# sql3Adpt.sqliteCursor.execute('''CREATE TABLE stocks
+#              (ID INTEGER PRIMARY KEY AUTOINCREMENT,name varchar(100),
+# count INT, price INT, reserved INT DEFAULT 0)''')
 
-sql3Adpt.sqliteCursor.execute("CREATE UNIQUE INDEX stoks_ind ON stocks (id)")
+# sql3Adpt.sqliteCursor.execute("CREATE UNIQUE INDEX stoks_ind ON stocks (id)")
 
 # Insert a row of data
 db.add_row("stocks",["product", 3, 100])
@@ -37,10 +38,10 @@ for row in db.select_all("stocks"):
 
 db.delete("stocks", [2,3] )
 
-sql3Adpt.sqliteCursor.execute("DROP TABLE stocks")
+# sql3Adpt.sqliteCursor.execute("DROP TABLE stocks")
 
-sql3Adpt.sqliteConnection.commit()
+sql3Adpt.sqlite_connection.commit()
 
 # We can also close the connection if we are done with it.
 # Just be sure any changes have been committed or they will be lost.
-sql3Adpt.sqliteConnection.close()
+sql3Adpt.sqlite_connection.close()
