--- ui/rus_locale.py
@@ -1,11 +1,0 @@
-# tab labels
-shop_tab = 'Магазин'
-customer_tab = 'Покупатели'
-
-# list column labels
-col = ['ID', 'Название товара', 'Количество', 'Цена', 'Зарезервировано']
-
-# button labels
-add_button = 'Добавить наименование'
-delete_button = 'Удалить'
-to_cart_button = 'Добавить в корзину'
